refactor: lean naming
