--- src/components/axfr-check/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "axrf-online-check.jsx"
-}
