--- src/components/axfr-domain-check/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "axfr-online-domain-check.jsx"
-}
