--- src/components/axfr-ssl-check/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "axfr-online-ssl-check.jsx"
-}
