--- src/components/circle-loading-button/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "circle-loading-button.jsx"
-}
