+++ src/components/side-panel-buttons/index.jsx
@@ -6,7 +6,7 @@ import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import { Link } from 'react-router-dom';
 
-import styles from './side-panel-buttons-styles';
+import styles from './styles';
 
 type Props = {
   classes: any,
