--- src/components/side-panel-buttons/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "side-panel-buttons.jsx"
-}
