+++ src/components/side-panel-buttons/styles.js
