--- src/containers/about-panel/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "about-panel.jsx"
-}
