--- src/containers/axfr-panel/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "axfr-panel.jsx"
-}
