--- src/containers/side-panel/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "side-panel.js"
-}
